rename modal to fix case issue
