+++ components/index.ts
@@ -1,4 +1,4 @@
 export { default as Event } from './event'
-export { default as Modal } from './modal'
+export { default as Modal } from './modal-what'
 export { default as Status } from './status'
 export { default as VideoPlayerStream } from './video-player-stream'
